+++ examples/ai-core/src/generate-image/black-forest-labs.ts
